初始化公司债收益率；全局变量文件拆分；定时任务非交易日不执行
