+++ datacenter/datacenter.go
@@ -2,6 +2,7 @@
 package datacenter
 
 import (
+	"github.com/axiaoxin-com/investool/datacenter/chinabond"
 	"github.com/axiaoxin-com/investool/datacenter/eastmoney"
 	"github.com/axiaoxin-com/investool/datacenter/eniu"
 	"github.com/axiaoxin-com/investool/datacenter/sina"
@@ -17,6 +18,8 @@ var (
 	Sina sina.Sina
 	// Zszx 招商证券
 	Zszx zszx.Zszx
+	// ChinaBond 中国债券信息网
+	ChinaBond chinabond.ChinaBond
 )
 
 func init() {
@@ -24,4 +27,5 @@ func init() {
 	Eniu = eniu.NewEniu()
 	Sina = sina.NewSina()
 	Zszx = zszx.NewZszx()
+	ChinaBond = chinabond.NewChinaBond()
 }
